+++ src/util/format.js
@@ -5,6 +5,7 @@ import { nb as locale } from 'date-fns/locale';
 const o = { locale };
 
 export const dateLong = (d) => format(d, 'PPP', o);
+export const dateLongCompact = (d) => format(d, 'PP', o);
 export const dateShort = (d) => format(d, 'P', o);
 export const dateTimeShort = (d) => format(d, 'Pp', o);
 export const dateTimeLong = (d) => format(d, 'PPPp', o);